Don't convert assign op operand types in opcache
This is the same change as 56b18d478ed9e40afd66860e82017d5c2017eac1
but for ASSIGN_OP. Changing the operand type may change the error
message and can result in different behavior with operator overloading.

As with the other patch, if there is strong interest this could be
added to the DFA pass instead, with an appropriate type check.